Restructured code and renamed LocationViewModel.kt to MainActivityViewModel.kt
