+++ app/src/main/java/knightec/bicyclepriority/presentation/repository/LocationData.kt
@@ -1,4 +1,4 @@
-package knightec.bicyclepriority.presentation
+package knightec.bicyclepriority.presentation.repository
 
 import android.Manifest
 import android.content.Context
@@ -9,6 +9,7 @@ import android.widget.Toast
 import androidx.core.app.ActivityCompat
 import androidx.lifecycle.LiveData
 import com.google.android.gms.location.*
+import knightec.bicyclepriority.presentation.LocationDetails
 
 class LocationData (var context : Context) : LiveData<LocationDetails>() {
     private val fusedLocationClient = LocationServices.getFusedLocationProviderClient(context)
